requirements.txt
django==3.2.2
djangorestframework==3.12.4
django-filter==21.1
drf-yasg==1.20
jsonschema==3.2.0
swagger-spec-validator==2.7.3
python-decouple==3.4
psycopg2-binary==2.9.1
pytest==6.2.5
pytest-django==4.4.0